+++ Runtime/NetworkDriver.cs
@@ -907,7 +907,9 @@ namespace Unity.Networking.Transport
             m_PipelineProcessor.UpdateReceive(this, out var updateCount);
 
             // TODO: Find a good way to establish a good limit (connections*pipelines/2?)
-            if (updateCount > (m_ConnectionList.Length - m_FreeList.Count) * 64)
+            var updateLimit = math.max(0, (m_ConnectionList.Length - m_FreeList.Count) * 64);
+
+            if (updateCount > updateLimit)
             {
                 UnityEngine.Debug.LogWarning(
                     FixedString.Format("A lot of pipeline updates have been queued, possibly too many being scheduled in pipeline logic, queue count: {0}", updateCount));
@@ -915,7 +917,7 @@ namespace Unity.Networking.Transport
 
             m_DefaultHeaderFlags = UdpCHeader.HeaderFlags.HasPipeline;
             m_PipelineProcessor.UpdateSend(ToConcurrentSendOnly(), out updateCount);
-            if (updateCount > (m_ConnectionList.Length - m_FreeList.Count) * 64)
+            if (updateCount > updateLimit)
             {
                 UnityEngine.Debug.LogWarning(
                     FixedString.Format("A lot of pipeline updates have been queued, possibly too many being scheduled in pipeline logic, queue count: {0}", updateCount));