+++ Runtime/Relay/RelayConnectionData.cs
@@ -28,12 +28,12 @@ namespace Unity.Networking.Transport.Relay
         /// <returns>Returns a RelayConnectionData constructed from the provided data.</returns>
         public static RelayConnectionData FromBytePointer(byte* dataPtr, int length)
         {
-            if (length != k_Length)
+            if (length > k_Length)
             {
 #if ENABLE_UNITY_COLLECTIONS_CHECKS
-                throw new ArgumentException($"Provided byte array length is invalid, must be {k_Length} but got {length}.");
+                throw new ArgumentException($"Provided byte array length is invalid, must be less or equal to {k_Length} but got {length}.");
 #else
-                UnityEngine.Debug.LogError($"Provided byte array length is invalid, must be {k_Length} but got {length}.");
+                UnityEngine.Debug.LogError($"Provided byte array length is invalid, must be less or equal to {k_Length} but got {length}.");
                 return default;
 #endif
             }
@@ -41,6 +41,17 @@ namespace Unity.Networking.Transport.Relay
             var connectionData = new RelayConnectionData();
             UnsafeUtility.MemCpy(connectionData.Value, dataPtr, length);
             return connectionData;
+        }
+
+        /// <summary>Convert a byte array to a <see cref="RelayConnectionData"/></summary>
+        /// <param name="data">Array to convert.</param>
+        /// <returns>New <see cref="RelayConnectionData"/>.</returns>
+        public static RelayConnectionData FromByteArray(byte[] data)
+        {
+            fixed (byte* ptr = data)
+            {
+                return RelayConnectionData.FromBytePointer(ptr, data.Length);
+            }
         }
     }
 }