+++ Runtime/Relay/RelayServerData.cs
@@ -1,7 +1,12 @@
 using System;
+using System.Linq;
 using System.Diagnostics;
 using Unity.Collections;
 using Unity.Collections.LowLevel.Unsafe;
+
+#if RELAY_SDK_INSTALLED
+using Unity.Services.Relay.Models;
+#endif
 
 namespace Unity.Networking.Transport.Relay
 {
@@ -43,6 +48,81 @@ namespace Unity.Networking.Transport.Relay
         /// </summary>
         public readonly byte IsSecure;
 
+        // Common code of all byte array-based constructors.
+        private RelayServerData(byte[] allocationId, byte[] connectionData, byte[] hostConnectionData, byte[] key)
+        {
+            Nonce = 0;
+            AllocationId = RelayAllocationId.FromByteArray(allocationId);
+            ConnectionData = RelayConnectionData.FromByteArray(connectionData);
+            HostConnectionData = RelayConnectionData.FromByteArray(hostConnectionData);
+            HMACKey = RelayHMACKey.FromByteArray(key);
+
+            // Assign temporary values to those. Chained constructors will set them.
+            Endpoint = default;
+            IsSecure = 0;
+
+            fixed(byte* hmacPtr = HMAC)
+            {
+                ComputeBindHMAC(hmacPtr, Nonce, ref ConnectionData, ref HMACKey);
+            }
+        }
+
+#if RELAY_SDK_INSTALLED
+        /// <summary>Create a new Relay server data structure from an allocation.</summary>
+        /// <param name="allocation">Allocation from which to create the server data.</param>
+        /// <param name="connectionType">Type of connection to use ("udp", "dtls", "ws", or "wss").</param>
+        public RelayServerData(Allocation allocation, string connectionType)
+            : this(allocation.AllocationIdBytes, allocation.ConnectionData, allocation.ConnectionData, allocation.Key)
+        {
+            // We check against a hardcoded list of strings instead of just trying to find the
+            // connection type in the endpoints since it may contains things we don't support
+            // (e.g. they provide a "tcp" endpoint which we don't support).
+            var supportedConnectionTypes = new string[] { "udp", "dtls" };
+            if (!supportedConnectionTypes.Contains(connectionType))
+                throw new ArgumentException($"Invalid connection type: {connectionType}. Must be udp or dtls.");
+
+            var serverEndpoint = allocation.ServerEndpoints.First(ep => ep.ConnectionType == connectionType);
+
+            Endpoint = HostToEndpoint(serverEndpoint.Host, (ushort)serverEndpoint.Port);
+            IsSecure = serverEndpoint.Secure ? (byte)1 : (byte)0;
+        }
+
+        /// <summary>Create a new Relay server data structure from a join allocation.</summary>
+        /// <param name="allocation">Allocation from which to create the server data.</param>
+        /// <param name="connectionType">Type of connection to use ("udp", "dtls", "ws", or "wss").</param>
+        public RelayServerData(JoinAllocation allocation, string connectionType)
+            : this(allocation.AllocationIdBytes, allocation.ConnectionData, allocation.HostConnectionData, allocation.Key)
+        {
+            // We check against a hardcoded list of strings instead of just trying to find the
+            // connection type in the endpoints since it may contains things we don't support
+            // (e.g. they provide a "tcp" endpoint which we don't support).
+            var supportedConnectionTypes = new string[] { "udp", "dtls" };
+            if (!supportedConnectionTypes.Contains(connectionType))
+                throw new ArgumentException($"Invalid connection type: {connectionType}. Must be udp, or dtls.");
+
+            var serverEndpoint = allocation.ServerEndpoints.First(ep => ep.ConnectionType == connectionType);
+
+            Endpoint = HostToEndpoint(serverEndpoint.Host, (ushort)serverEndpoint.Port);
+            IsSecure = serverEndpoint.Secure ? (byte)1 : (byte)0;
+        }
+#endif
+
+        /// <summary>Create a new Relay server data structure.</summary>
+        /// <param name="host">IP address of the Relay server.</param>
+        /// <param name="port">Port of the Relay server.</param>
+        /// <param name="allocationId">ID of the Relay allocation.</param>
+        /// <param name="connectionData">Connection data of the allocation.</param>
+        /// <param name="hostConnectionData">Connection data of the host (same as previous for hosts).</param>
+        /// <param name="key">HMAC signature of the allocation.</param>
+        /// <param name="isSecure">Whether the Relay connection is to be secured or not.</param>
+        public RelayServerData(string host, ushort port, byte[] allocationId, byte[] connectionData,
+            byte[] hostConnectionData, byte[] key, bool isSecure)
+            : this(allocationId, connectionData, hostConnectionData, key)
+        {
+            Endpoint = HostToEndpoint(host, port);
+            IsSecure = isSecure ? (byte)1 : (byte)0;
+        }
+
         /// <summary>
         /// Initializes a new instance of the <see cref="RelayServerData"/> class
         /// </summary>
@@ -53,6 +133,7 @@ namespace Unity.Networking.Transport.Relay
         /// <param name="hostConnectionData">The host connection data</param>
         /// <param name="key">The key</param>
         /// <param name="isSecure">The is secure</param>
+        [Obsolete("Will be removed in Unity Transport 2.0. Use the new constructor introduced in 1.3 instead.", false)]
         public RelayServerData(ref NetworkEndPoint endpoint, ushort nonce, RelayAllocationId allocationId, string connectionData, string hostConnectionData, string key, bool isSecure)
         {
             Endpoint = endpoint;
@@ -107,6 +188,7 @@ namespace Unity.Networking.Transport.Relay
         /// <summary>
         /// Computes the new nonce, this must be called one time!
         /// </summary>
+        [Obsolete("Will be removed in Unity Transport 2.0. There shouldn't be any need to call this method.")]
         public void ComputeNewNonce()
         {
             Nonce = (ushort)(new Unity.Mathematics.Random((uint)Stopwatch.GetTimestamp())).NextUInt(1, 0xefff);
@@ -153,6 +235,20 @@ namespace Unity.Networking.Transport.Relay
 
                 HMACSHA256.ComputeHash(keyValue, keyArray.Length, messageBytes, messageLength, result);
             }
+        }
+
+        private static NetworkEndPoint HostToEndpoint(string host, ushort port)
+        {
+            NetworkEndPoint endpoint;
+
+            if (NetworkEndPoint.TryParse(host, port, out endpoint, NetworkFamily.Ipv4))
+                return endpoint;
+
+            if (NetworkEndPoint.TryParse(host, port, out endpoint, NetworkFamily.Ipv6))
+                return endpoint;
+
+            UnityEngine.Debug.LogError($"Host {host} is not a valid IPv4 or IPv6 address.");
+            return endpoint;
         }
     }
 }